# PathKeeper/backend-node/src/openapi.yaml
openapi: 3.0.3
info:
  title: PathKeepers Backend API
  version: 0.1.0
  description: >-
    Early scaffold of PathKeepers API. This spec will expand as features evolve.
servers:
  - url: http://localhost:6060
paths:
...
  /api/auth/login:
    post:
      summary: Login (temporary scaffold)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Auth success
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  token: { type: string }
                  user:
                    type: object
                    properties:
                      id: { type: string }
                      email: { type: string }
                      role: { type: string }
                      name: { type: string }
        '401':
          description: Invalid credentials
  /api/auth/me:
    get:
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User info
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT